pyproject.toml
[project]
name = "a2a-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Pierre MELLET",email = "pierre.mellet@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

[tool.poetry]

[tool.poetry.dependencies]
langchain = "^0.3.23"
langchain-openai = "^0.3.12"
langgraph = "^0.3.28"
jwcrypto = "^1.5.6"
python-dotenv = "^1.1.0"
debugpy = "^1.8.14"
httpx-sse = "^0.4.0"
sse-starlette = "^2.2.1"
pyjwt = "^2.10.1"
chainlit = "^2.5.5"
termcolor = "^3.0.1"
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
